+++ livedata/src/main/java/io/getstream/chat/android/livedata/controller/ChannelControllerImpl.kt
@@ -509,27 +509,22 @@ class ChannelControllerImpl(
         val mimeType = MimeTypeMap.getSingleton().getMimeTypeFromExtension(file.extension)
         val attachmentType = if (mimeType.isImageMimetype()) { "image" } else { "file" }
         val pathResult = if (attachmentType == "image") { sendImage(file) } else { sendFile(file) }
-        var newAttachment: Attachment
-        var uploadError: ChatError? = null
+        val url = if (pathResult.isError) null else pathResult.data()
+        val uploadState = if (pathResult.isError) Attachment.UploadState.Failed(pathResult.error()) else Attachment.UploadState.Success
 
-        if (pathResult.isError) {
-            uploadError = pathResult.error()
-
-            newAttachment = attachment.copy(uploadState = Attachment.UploadState.Failed(uploadError))
-        } else {
-            val uploadPath = pathResult.data()
-            newAttachment = attachment.copy(
-                name = file.name,
-                fileSize = file.length().toInt(),
-                mimeType = mimeType?.toString() ?: "",
-                url = uploadPath,
-                uploadState = Attachment.UploadState.Success,
-                type = attachmentType
-            ).apply {
+        var newAttachment = attachment.copy(
+            name = file.name,
+            fileSize = file.length().toInt(),
+            mimeType = mimeType ?: "",
+            url = url,
+            uploadState = uploadState,
+            type = attachmentType
+        ).apply {
+            url?.let {
                 if (attachmentType == "image") {
-                    imageUrl = uploadPath
+                    imageUrl = it
                 } else {
-                    assetUrl = uploadPath
+                    assetUrl = it
                 }
             }
         }
@@ -539,7 +534,7 @@ class ChannelControllerImpl(
             newAttachment = attachmentTransformer(newAttachment, file)
         }
 
-        return Result(newAttachment, uploadError)
+        return Result(newAttachment, if (pathResult.isError) pathResult.error() else null)
     }
 
     /**